[jb-contrib] Merged trunk (and apprently restored VC++08 files)

+++ branches/jb-contrib/MTA10_Server/mods/deathmatch/logic/CPerfStat.BandwidthUsage.cpp
@@ -258,7 +258,7 @@ void CPerfStatBandwidthUsageImpl::LoadStats ( void )
         for ( int r = 0 ; r < result.nRows ; r++ )
         {
             SString strType = (const char*)result.Data[r][0].pVal;
-            uint uiIndex = result.Data[r][1].nVal;
+            uint uiIndex = static_cast < uint > ( result.Data[r][1].nVal );
             float GameRecv = Max ( 0.f, result.Data[r][2].fVal );
             float GameSent = Max ( 0.f, result.Data[r][3].fVal );
             float HttpSent = Max ( 0.f, result.Data[r][4].fVal );
